pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "relatorio-pedidos-spark"
version = "0.1.0"
description = "Relatório PySpark: pedidos com pagamento recusado e fraude legítima (FIAP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyspark>=3.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
